15: Implement unit tests.
`make test`
Configure Dockerfile to run tests in container: `docker compose run --rm src test`.
