+++ Makefile
@@ -1,0 +1,6 @@
+.PHONY: default
+default: test
+
+.PHONY: test
+test:
+	./vendor/bin/phpunit --testdox src/tests
