+++ composer.json
@@ -1,0 +1,10 @@
+{
+    "autoload": {
+        "classmap": [
+            "src/include/"
+        ]
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^9.5"
+    }
+}
