+++ docker-compose.yaml
@@ -1,0 +1,7 @@
+version: "3.5"
+services:
+  src:
+    build: .
+    volumes:
+      - ./src:/build/src
+      - ./assets:/build/assets
